fix: prevent form from being disabled when submitting
This commit addresses an issue where the form submit button was being disabled prematurely when the form was submitted, even if the form was valid.

The `disabled` prop on the `FormButton` component was commented out to prevent this behavior. This allows the form to be submitted regardless of the `isValid` and `isSubmitted` states.